extracted files to own files and moved to folders
