+++ src/QueryPressure.UI/Services/ExecutionStore.cs
@@ -1,32 +1,11 @@
 using System.Collections.Concurrent;
 using System.Collections.Immutable;
-using System.Collections.ObjectModel;
-using System.Threading.Channels;
 using QueryPressure.App.Interfaces;
 using QueryPressure.Core.Interfaces;
+using QueryPressure.UI.Inderfaces;
+using QueryPressure.UI.Models;
 
-namespace QueryPressure.UI;
-
-public record Execution(Guid Id,
-  Task ExecutionTask,
-  IExecutionResultStore ExecutionResultStore,
-  IReadOnlyCollection<ILiveMetricProvider> MetricProviders,
-  CancellationTokenSource CancellationTokenSource);
-
-public interface IExecutionStore
-{
-  ImmutableArray<Execution> RunningExecutions { get; }
-  ImmutableArray<Execution> CompletedExecutions { get; }
-  
-  Task<Guid> SaveAsync(Task executionTask,
-    IExecutionResultStore executionResultStore,
-    IReadOnlyCollection<ILiveMetricProvider> liveMetricProviders,
-    CancellationTokenSource cts);
-
-  Task CancelAsync(Guid executionId); 
-  
-  Task RemoveAsync(Guid executionId);
-}
+namespace QueryPressure.UI.Services;
 
 public class ExecutionStore : IExecutionStore
 {
